chore(api): modularize content generation DTOs and schemas
- Extracted MessageDto and WorkflowContextDto into separate files for better organization and reusability.
- Updated GenerateContentRequestDto to utilize the new DTOs, enhancing clarity and maintainability.
- Introduced content generation schemas for various notification types, streamlining the structure of generated content.
- Added constants for content generation guidelines, improving the overall content generation process.
