+++ apps/api/src/app/ai/dtos/generate-content-request.dto.ts
@@ -1,53 +1,9 @@
 import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
 import { StepTypeEnum } from '@novu/shared';
 import { Type } from 'class-transformer';
-import { IsArray, IsEnum, IsNotEmpty, IsObject, IsOptional, IsString, ValidateNested } from 'class-validator';
-
-export enum MessageRole {
-  USER = 'user',
-  ASSISTANT = 'assistant',
-}
-
-export class MessageDto {
-  @ApiProperty({
-    description: 'Role of the message sender',
-    enum: MessageRole,
-  })
-  @IsEnum(MessageRole)
-  role: MessageRole;
-
-  @ApiProperty({
-    description: 'Content of the message',
-  })
-  @IsString()
-  @IsNotEmpty()
-  content: string;
-}
-
-export class WorkflowContextDto {
-  @ApiPropertyOptional({
-    description: 'Name of the workflow',
-  })
-  @IsOptional()
-  @IsString()
-  workflowName?: string;
-
-  @ApiPropertyOptional({
-    description: 'Description of the workflow',
-  })
-  @IsOptional()
-  @IsString()
-  workflowDescription?: string;
-
-  @ApiPropertyOptional({
-    description: 'Available variables in the step',
-    type: [String],
-  })
-  @IsOptional()
-  @IsArray()
-  @IsString({ each: true })
-  variables?: string[];
-}
+import { IsArray, IsEnum, IsObject, IsOptional, IsString, ValidateNested } from 'class-validator';
+import { MessageDto } from './message.dto';
+import { WorkflowContextDto } from './workflow-context.dto';
 
 export class GenerateContentRequestDto {
   @ApiProperty({
