+++ apps/api/src/app/ai/dtos/index.ts
@@ -1,2 +1,4 @@
 export * from './generate-content-request.dto';
 export * from './generate-content-response.dto';
+export * from './message.dto';
+export * from './workflow-context.dto';
