+++ apps/api/src/app/ai/usecases/generate-content/generate-content.constants.ts
@@ -1,0 +1,71 @@
+import { StepTypeEnum } from '@novu/shared';
+
+export const BASE_SYSTEM_PROMPT = `You are a notification content expert for Novu, a notification infrastructure platform.
+Your task is to generate high-quality notification content based on user requests.
+
+Guidelines:
+- Be concise and engaging
+- Use professional but friendly tone
+- Include personalization placeholders where appropriate using Liquid syntax: {{subscriber.firstName}}, {{payload.variableName}}
+- Follow best practices for the specific notification channel`;
+
+export const EMAIL_HTML_GUIDELINES = `
+Email-specific guidelines (HTML):
+- Subject lines should be compelling and under 60 characters
+- Body must be valid HTML with inline styles for email client compatibility
+- Use semantic HTML: <h1>, <h2>, <p>, <a>, <table> for layout
+- Add inline styles for colors, spacing, fonts (e.g., style="color: #333; margin: 16px 0;")
+- Include variables using Liquid syntax: {{subscriber.firstName}}, {{payload.variableName}}
+- Use tables for layout to ensure compatibility across email clients
+- Keep paragraphs short and scannable
+- Use clear call-to-action links or buttons
+- Structure: greeting -> main content -> CTA -> closing`;
+
+export const EMAIL_BLOCK_GUIDELINES = `
+Email-specific guidelines (Block Editor):
+- Subject lines should be compelling and under 60 characters
+- Body must be in Maily TipTap JSON format with proper node structure
+- Use heading nodes for titles (level 1 for main, 2 for sections)
+- Use paragraph nodes for body text
+- Use spacer nodes between sections (height: 16 or 24)
+- Use button nodes for CTAs with good contrast colors
+- Include variables using variable nodes with id like "subscriber.firstName" or "payload.variableName"
+- Keep paragraphs short and scannable
+- Structure: greeting -> main content -> CTA button -> closing`;
+
+export const SMS_GUIDELINES = `
+SMS-specific guidelines:
+- Keep messages under 160 characters to avoid splitting
+- Be direct and actionable
+- Include essential information only
+- Avoid special characters that might not render properly`;
+
+export const PUSH_GUIDELINES = `
+Push notification guidelines:
+- Title should be under 50 characters
+- Body should be under 150 characters
+- Create urgency or value proposition
+- Be specific about what the user should do`;
+
+export const IN_APP_GUIDELINES = `
+In-app notification guidelines:
+- Can be slightly longer than push notifications
+- Include action buttons when appropriate
+- Be contextual to the user's current state
+- Use clear, actionable language`;
+
+export const CHAT_GUIDELINES = `
+Chat message guidelines:
+- Keep messages conversational
+- Be friendly but professional
+- Include relevant context
+- Make it easy to respond or take action`;
+
+export const CHANNEL_GUIDELINES: Partial<Record<StepTypeEnum, string | ((editorType?: string) => string)>> = {
+  [StepTypeEnum.EMAIL]: (editorType?: string) =>
+    editorType === 'html' ? EMAIL_HTML_GUIDELINES : EMAIL_BLOCK_GUIDELINES,
+  [StepTypeEnum.SMS]: SMS_GUIDELINES,
+  [StepTypeEnum.PUSH]: PUSH_GUIDELINES,
+  [StepTypeEnum.IN_APP]: IN_APP_GUIDELINES,
+  [StepTypeEnum.CHAT]: CHAT_GUIDELINES,
+};
