+++ apps/api/src/app/ai/usecases/generate-content/generate-content.schemas.ts
@@ -1,0 +1,175 @@
+import { z } from 'zod';
+
+/**
+ * AI Content Generation Schemas
+ *
+ * These schemas define the structure of content generated by AI for each channel type.
+ * They are focused on content only (subject, body, actions) and do not include
+ * configuration fields like 'skip', 'layoutId', etc.
+ *
+ * Related schemas in @novu/application-generic:
+ * - emailControlZodSchema: Full email step configuration (includes subject, body, layoutId, skip)
+ * - smsControlZodSchema: Full SMS step configuration (includes body, skip)
+ * - pushControlZodSchema: Full push step configuration (includes subject, body, skip)
+ * - inAppControlZodSchema: Full in-app step configuration (includes subject, body, actions, redirect, skip)
+ * - chatControlZodSchema: Full chat step configuration (includes body, skip)
+ */
+
+/**
+ * Maily TipTap JSON Schema
+ * Defines the structure for email block editor content.
+ * These schemas represent the TipTap JSON format used by Maily for rich email editing.
+ */
+const mailyTextNodeSchema = z.object({
+  type: z.literal('text'),
+  text: z.string(),
+  marks: z
+    .array(
+      z.object({
+        type: z.enum(['bold', 'italic', 'underline', 'strike']),
+      })
+    )
+    .optional(),
+});
+
+const mailyVariableNodeSchema = z.object({
+  type: z.literal('variable'),
+  attrs: z.object({
+    id: z.string().describe('Variable name like subscriber.firstName or payload.companyName'),
+    fallback: z.string().nullable().optional(),
+  }),
+});
+
+const mailyParagraphSchema = z.object({
+  type: z.literal('paragraph'),
+  attrs: z.object({ textAlign: z.enum(['left', 'center', 'right']).nullable().optional() }).optional(),
+  content: z.array(z.union([mailyTextNodeSchema, mailyVariableNodeSchema])).optional(),
+});
+
+const mailyHeadingSchema = z.object({
+  type: z.literal('heading'),
+  attrs: z.object({
+    level: z
+      .union([z.literal(1), z.literal(2), z.literal(3)])
+      .describe('Heading level: 1 for main title, 2 for section, 3 for subsection'),
+    textAlign: z.enum(['left', 'center', 'right']).nullable().optional(),
+  }),
+  content: z.array(z.union([mailyTextNodeSchema, mailyVariableNodeSchema])).optional(),
+});
+
+const mailyButtonSchema = z.object({
+  type: z.literal('button'),
+  attrs: z.object({
+    text: z.string().describe('Button label text'),
+    url: z.string().optional().describe('Button link URL'),
+    isTextVariable: z.literal(false).optional(),
+    isUrlVariable: z.literal(false).optional(),
+    alignment: z.enum(['left', 'center', 'right']).optional(),
+    variant: z.enum(['filled', 'outline']).optional(),
+    borderRadius: z.enum(['smooth', 'sharp', 'round']).optional(),
+    buttonColor: z.string().optional().describe('Hex color like #000000'),
+    textColor: z.string().optional().describe('Hex color like #ffffff'),
+  }),
+});
+
+const mailySpacerSchema = z.object({
+  type: z.literal('spacer'),
+  attrs: z.object({
+    height: z.number().describe('Height in pixels, typically 8, 16, 24, or 32'),
+  }),
+});
+
+const mailyDividerSchema = z.object({
+  type: z.literal('horizontalRule'),
+});
+
+const mailyNodeSchema = z.union([
+  mailyParagraphSchema,
+  mailyHeadingSchema,
+  mailyButtonSchema,
+  mailySpacerSchema,
+  mailyDividerSchema,
+]);
+
+const mailyBodySchema = z.object({
+  type: z.literal('doc'),
+  content: z.array(mailyNodeSchema).describe('Array of content nodes that make up the email body'),
+});
+
+/**
+ * Email content schemas for AI generation
+ * Block editor uses Maily TipTap JSON, HTML editor uses raw HTML string
+ */
+export const emailBlockContentSchema = z.object({
+  subject: z.string().describe('Email subject line - concise and engaging, under 60 characters'),
+  body: mailyBodySchema.describe('Email body in Maily TipTap JSON format'),
+});
+
+export const emailHtmlContentSchema = z.object({
+  subject: z.string().describe('Email subject line - concise and engaging, under 60 characters'),
+  body: z.string().describe('Email body in HTML format - use semantic HTML with inline styles'),
+});
+
+/**
+ * SMS content schema for AI generation
+ * Derived from smsControlZodSchema but focused on content only
+ */
+export const smsContentSchema = z.object({
+  body: z.string().max(160).describe('SMS message body - keep under 160 characters'),
+});
+
+/**
+ * Push content schema for AI generation
+ * Derived from pushControlZodSchema but focused on content only
+ */
+export const pushContentSchema = z.object({
+  subject: z.string().max(50).describe('Push notification title - keep under 50 characters'),
+  body: z.string().max(150).describe('Push notification body - keep under 150 characters'),
+});
+
+/**
+ * In-app content schema for AI generation
+ * Simplified version of inAppControlZodSchema focused on content.
+ * Note: inAppControlZodSchema includes redirect, avatar, data, and skip fields
+ * which are not needed for AI content generation.
+ */
+export const inAppContentSchema = z.object({
+  subject: z.string().optional().describe('In-app notification subject/title'),
+  body: z.string().describe('In-app notification body content'),
+  primaryAction: z
+    .object({
+      label: z.string().describe('Primary button label'),
+      url: z.string().optional().describe('URL to navigate to when clicked'),
+    })
+    .optional()
+    .describe('Primary action button'),
+  secondaryAction: z
+    .object({
+      label: z.string().describe('Secondary button label'),
+      url: z.string().optional().describe('URL to navigate to when clicked'),
+    })
+    .optional()
+    .describe('Secondary action button'),
+});
+
+/**
+ * Chat content schema for AI generation
+ * Derived from chatControlZodSchema but focused on content only
+ */
+export const chatContentSchema = z.object({
+  body: z.string().describe('Chat message content'),
+});
+
+export const responseWrapperSchema = (contentSchema: z.ZodTypeAny) =>
+  z.object({
+    aiMessage: z
+      .string()
+      .describe('A brief, friendly message explaining what you generated and any suggestions for the user'),
+    content: contentSchema,
+    suggestedPayload: z
+      .record(z.string(), z.string())
+      .optional()
+      .describe(
+        'Sample values ONLY for {{payload.*}} variables used in the content. Keys should be the variable name WITHOUT the "payload." prefix (e.g., for {{payload.link}} use key "link"). NEVER include subscriber variables here - subscriber.firstName, subscriber.email etc are provided by the system.'
+      ),
+  });
