+++ apps/api/src/app/ai/usecases/generate-content/index.ts
@@ -1,2 +1,4 @@
 export * from './generate-content.command';
+export * from './generate-content.constants';
+export * from './generate-content.schemas';
 export * from './generate-content.usecase';
